signing-storage-urls/requirements.txt
Flask==1.1.2
google-auth==1.24.0
google-cloud-storage==1.34.0
